+++ benchmark/a2c/atari_model.py
@@ -2,7 +2,7 @@
 Author: jianzhnie
 Date: 2022-09-02 14:36:16
 LastEditors: jianzhnie
-LastEditTime: 2022-09-02 14:36:19
+LastEditTime: 2022-09-02 19:48:05
 Description:
 Copyright (c) 2022 by jianzhnie@126.com, All Rights Reserved.
 '''
@@ -11,7 +11,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from rltoolkit.core.model import Model
+from rltoolkit.core.models.base_model import Model
 
 
 class ActorCritic(Model):
