+++ benchmark/quickstart/cartpole_model.py
@@ -2,7 +2,7 @@
 Author: jianzhnie
 Date: 2022-09-02 12:19:54
 LastEditors: jianzhnie
-LastEditTime: 2022-09-02 12:21:11
+LastEditTime: 2022-09-02 19:47:59
 Description:
 Copyright (c) 2022 by jianzhnie@126.com, All Rights Reserved.
 '''
@@ -10,7 +10,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from rltoolkit.core.model import Model
+from rltoolkit.core.models.base_model import Model
 
 
 class CartpoleModel(Model):
